+++ gcovr/formats/cobertura/read.py
@@ -87,12 +87,12 @@ def read_report(options: Options) -> CovData:
             if is_file_excluded(file_path, options.filter, options.exclude):
                 continue
 
-            file_coverage = FileCoverage(file_path, filename)
+            filecov = FileCoverage(file_path, filename)
             merge_options = get_merge_mode_from_options(options)
             for xml_line in gcovr_file.xpath("./lines//line"):
-                insert_line_coverage(file_coverage, _line_from_xml(filename, xml_line))
+                insert_line_coverage(filecov, _line_from_xml(filename, xml_line))
 
-            insert_file_coverage(covdata, file_coverage, merge_options)
+            insert_file_coverage(covdata, filecov, merge_options)
 
     return covdata