+++ gcovr/formats/txt/write.py
@@ -179,7 +179,7 @@ def _format_line(name: str, stat: CoverageStat, uncovered_lines: str) -> str:
 
 def _covered_lines_str(filecov: FileCoverage) -> str:
     covered_lines = sorted(
-        line.lineno for line in filecov.lines.values() if not line.is_uncovered
+        linecov.lineno for linecov in filecov.lines.values() if not linecov.is_uncovered
     )
 
     # Walk through the covered lines in sorted order.
@@ -197,7 +197,7 @@ def _covered_lines_str(filecov: FileCoverage) -> str:
 
 def _uncovered_lines_str(filecov: FileCoverage) -> str:
     uncovered_lines = sorted(
-        line.lineno for line in filecov.lines.values() if line.is_uncovered
+        linecov.lineno for linecov in filecov.lines.values() if linecov.is_uncovered
     )
 
     # Walk through the uncovered lines in sorted order.
@@ -215,7 +215,9 @@ def _uncovered_lines_str(filecov: FileCoverage) -> str:
 
 def _covered_branches_str(filecov: FileCoverage) -> str:
     covered_lines = sorted(
-        line.lineno for line in filecov.lines.values() if not line.has_uncovered_branch
+        linecov.lineno
+        for linecov in filecov.lines.values()
+        if not linecov.has_uncovered_branch
     )
 
     # Don't do any aggregation on branch results.
@@ -224,23 +226,27 @@ def _covered_branches_str(filecov: FileCoverage) -> str:
 
 def _covered_decisions_str(filecov: FileCoverage) -> str:
     covered_decisions = sorted(
-        line.lineno
-        for line in filecov.lines.values()
-        if not line.has_uncovered_decision
+        linecov.lineno
+        for linecov in filecov.lines.values()
+        if not linecov.has_uncovered_decision
     )
     return ",".join(str(lineno) for lineno in covered_decisions)
 
 
 def _uncovered_decisions_str(filecov: FileCoverage) -> str:
     uncovered_decisions = sorted(
-        line.lineno for line in filecov.lines.values() if line.has_uncovered_decision
+        linecov.lineno
+        for linecov in filecov.lines.values()
+        if linecov.has_uncovered_decision
     )
     return ",".join(str(lineno) for lineno in uncovered_decisions)
 
 
 def _uncovered_branches_str(filecov: FileCoverage) -> str:
     uncovered_lines = sorted(
-        line.lineno for line in filecov.lines.values() if line.has_uncovered_branch
+        linecov.lineno
+        for linecov in filecov.lines.values()
+        if linecov.has_uncovered_branch
     )
 
     # Don't do any aggregation on branch results.